Made sure that the selected and selectable hydraulic boundary locations for piping use the same reference points when calculation the distance. Added tests: - To ensure that the distance of the SelectedHydraulicBoundaryLocation is the same as the item in GetSelectableHydraulicBoundaryLocations
Note: 
Piping has no property to expose the ReferenceLineIntersectionWorldPoint point to the user. As a result, the reference points that are used for calculating the SelectedHydraulicBoundaryLocation and the items within the GetSelectableHydraulicBoundaryLocations can only use the ReferenceLineIntersectionWorldPoint that is encapsulated within the surfaceline class: it is assumed that this property is used for the distance calculations of both properties. Aforementioned test should be sufficient to prove that both properties use the same reference points for their distance calculations. 

- Removed redundant parentheses 
Related to Issue [WTI-1056]


git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@7699 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 399aff980417c31a1d8ef06d626a617b3e671ae9
Former-commit-id: 889fdcf6948b6c832a44d5273ff453888416cfe3