Review suggestion - simplify tests
{qoute}
Looking at most of this unit test, there doesn't seem to be any need for ControlTester to reflect the encapsulated member. Think about the level of abstraction you are working and think of needing thinks like ControlTester as potential smells that the object you are working on is encapsulating too much or you as a dev are too much focussed on doing white-box testing. 
{quote}
Related to Issue [WTI-1091]


git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@8437 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 159916a320a71d293c597116f3206f303cc16f8f
Former-commit-id: 1b797a01335db5609e7ce3062975cca2e36a99d4